+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,6 +3,23 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v1.6.0
+------------------------------
+*September 30, 2021*
+
+### Added
+- Analytics service file to handle analytics events.
+- `f-analytics` plugin to handle GTM events.
+
+
+v1.5.0
+------------------------------
+*September 30, 2021*
+
+### Changed
+- Only call Geolocation endpoint when the service type is delivery
+
+
 v1.4.0
 ------------------------------
 *September 29, 2021*